requirements.txt
asgiref==3.9.0
django==4.2.23
djangorestframework==3.16.0
PyJWT==2.10.1
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
xcu==1.0.2
